worker/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
deprecation==2.1.0
exceptiongroup==1.2.2
frozenlist==1.5.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
multidict==6.2.0
packaging==24.2
postgrest==0.19.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.1
requests==2.32.3
requests-random-user-agent==2023.10.25
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
storage3==0.11.3
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.3
typing-extensions==4.13.0
urllib3==1.26.20
websockets==14.2
xmltodict==0.14.2
yarl==1.18.3